requirements.txt
Django==1.3.1
psycopg2==2.4.2
dj-database-url
python-twitter
python-dateutil
django-json-field
BeautifulSoup
